refactor(app): clear deprecation warnings with real migrations
Replace deprecated APIs instead of suppressing them:

- ListItem: move headlineContent to the trailing content lambda (18 sites).
- rememberModalBottomSheetState -> rememberBottomSheetState with
  SheetValue.Hidden (5 sites); remove redundant cast in DetailsViewModel.
- Crypto: drop EncryptedSharedPreferences/MasterKey (whole library
  deprecated, no first-party successor) for direct Android Keystore
  AES-256-GCM encryption of the token JSON in plain SharedPreferences.
- Media3: drop MediaSessionCompat/PlaybackStateCompat/MediaMetadataCompat
  and the custom action receiver; bridge the engine-agnostic
  PlaybackSessionController to a SimpleBasePlayer and back a Media3
  MediaSession, with the notification using MediaStyleNotificationHelper.

+++ android/app/src/main/java/com/crispy/tv/details/DetailsViewModel.kt
@@ -697,7 +697,7 @@ class DetailsViewModel internal constructor(
                             if (season == null || videos.isEmpty()) {
                                 null
                             } else {
-                                (season as Int) to videos.all { watchStates[it.id]?.isWatched == true }
+                                season to videos.all { watchStates[it.id]?.isWatched == true }
                             }
                         }.toMap()
